lab5/requirements.txt
psycopg2-binary==2.9.9
pydantic-settings==2.1.0